tech(tests): create mocks and test module touchpoints
Previously, the tests were calling the actual `gatsby-source-filesystem` module and generating files/images in the `.cache` directory. The immidate result of this was that adding additional tests using the same `imageUrl`s was causing the module to hit the cache and find it had already generated a file for that url and then subsequent assertions that expected a new file to be returned would fail.

The other impetus for these changes is the practice of not unit testing an underlying technology, but rather testing your touchpoints with that technology. The previous code was making assertions to ensure that `gatsby-source-filesystem` was doing its job, but that's arguably not the responsibility of this module. We should be testing whether our code is forming the expected arguments to the module and that our code which operates on the return value is working as expected.

To address both of these points, these changes mock out the `createRemoteFileNode` method of `gatsby-source-filesystem` to perform the bare minimum logic, but instead stand in as an opportunity to test the arguments our code is passing to it and test the logic we trigger on the return value.

Other changes to the test file were to extract the mocks for the `gatsbyNodeHelpers` object, so that we get a fresh copy for each test, since this was also an issue when adding addition tests that reused the same mock functions for subsequent calls.